Disable pointer-events for the label div
Without this, a quickly-moving cursor can reach the label div before it is repositioned by Leaflet.label, where the div will intercept the hover events and disable the label. Once the label is gone, the hover event goes to the map below, causing a mouseover event and the label to show. Repeat a few dozen times per second and you get an annoying flicker.