Bug fix: clickable and noHide are mutually exclusive
Setting the option offset:[0, 0] and clickable:true causes the label <div>
be just under the mouse, firing a 'mouseout' on the Path, making the Label
disappear and re-appear again resulting in flickering or no label at all.
This make 'clickable' and 'noHide' mutually exclusive. Also, in
noHide:false mode, clicking the label is not possible anyway as it
moves with the mouse.